[bigquery] Move `encodeStructToJSONString` function
