+++ clients/bigquery/bigquery.go
@@ -34,8 +34,8 @@ const (
 	describeNameCol               = "column_name"
 	describeTypeCol               = "data_type"
 	describeCommentCol            = "description"
-	// Storage Write API is limited to 10 MiB, subtract 50 KiB to account for request overhead.
-	maxRequestByteSize = (10 * 1024 * 1024) - (50 * 1024)
+	// Storage Write API is limited to 10 MB, let's start out conservative and use 80% of that.
+	maxRequestByteSize = 10_000_000 * .8
 )
 
 type Store struct {
