+++ clients/bigquery/storagewrite_test.go
@@ -171,3 +171,30 @@ func TestRowToMessage(t *testing.T) {
 		"cArray":         []any{"foo", "bar"},
 	}, result)
 }
+
+func TestEncodeStructToJSONString(t *testing.T) {
+	{
+		// Empty string:
+		result, err := encodeStructToJSONString("")
+		assert.NoError(t, err)
+		assert.Equal(t, "", result)
+	}
+	{
+		// Toasted string:
+		result, err := encodeStructToJSONString("__debezium_unavailable_value")
+		assert.NoError(t, err)
+		assert.Equal(t, `{"key":"__debezium_unavailable_value"}`, result)
+	}
+	{
+		// Map:
+		result, err := encodeStructToJSONString(map[string]any{"foo": "bar", "baz": 1234})
+		assert.NoError(t, err)
+		assert.Equal(t, `{"baz":1234,"foo":"bar"}`, result)
+	}
+	{
+		// Toasted map (should not happen):
+		result, err := encodeStructToJSONString(map[string]any{"__debezium_unavailable_value": "bar", "baz": 1234})
+		assert.NoError(t, err)
+		assert.Equal(t, `{"key":"__debezium_unavailable_value"}`, result)
+	}
+}
